Planar path graph with generic algebraic labels
The class template PlanarGraph encapsulates the fundamental operations
needed to create and regularize a graph geometrically embedded in the
plane. Such graph is created through the insertion of edges, which are
geometrically represented as Paths; vertices are created automatically
and clumped based on a numerical threshold.

The crucial member function provided by PlanarGraph is "regularize()".
This algorithm sanitizes the entire graph, removing certain degenerate
configurations. More precisely, it merges edges which overlap (fully or
partially), removes degenerate loops and reglues teardrop- and lasso-
shaped paths. Moreover, it ensures that the edge incidences around each
vertex are sorted by increasing departure heading, and if headings are
equal (tangent paths), by the eventual direction where the paths go.

The template parameter of PlanarGraph is an additional class serving as
an edge label: an instance of it is stored with each edge. These labels
are algebraic - they keep track of the fundamental operations performed
during regularization: edge reversal, merger and removal.

The functionality provided by PlanarGraph forms a common geometric core
required for the future boolean operations engine. Using different types
of algebraic labels, it will be possible to implement shape flattening,
binary and n-arry boolean operations, as well as path slicing/division.