Add a method to find coordinate extrema on a path
Add Path::extrema() which returns "witness" extrema on the path. These
witnesses are points on the path at which the X or Y coordinate attain
their minimum or maximum values (as such points are not always unique,
one example is returned for each of the min and max). In addition, for
each of the (min and max) witnesses, the method returns a PathTime and
tells us whether the other coordinate (not the one being maximized) is
increasing, decreasing or is critical at these path-times.