Improve numerical stability of self-intersections
This is a rewrite of the curve self-intersection algorithm implemented
in Curve::intersectSelf(). The main feature of the new implementation,
and the goal of this change, is the elimination of spurious crossings.
These spurious self-intersections were typically reported at critical
points of the X-coordinate, where the old algorithm cut the curve into
monotonic pieces. Due to numerical errors, these pieces were sometimes
wrongly reported to intersect at points near to the cut locations.

The new version essentially runs two passes of the original algorithm,
but uses X-coordinate in one pass and Y-coordinate in the other. Only
the self-intersections found by both passes are considered valid and
are returned to the user.

By ensuring cleaner output, this change reduces the need for self-
intersection filtering in the higher-level routines such as the Path
and PathVector self-intersection functions. In particular, the over-
filtering which removed real (valid) self-intersections is prevented,
fixing the issue in the reference. A separate problem with spurious
intersections near the juncture between two curves in a Path is also
solved.

New unit tests are added for self-intersection algorithms, including
a regression test for the fixed issue.

Fixes: https://gitlab.com/inkscape/lib2geom/-/issues/46