CI: Split off tests into a separate pipeline job
Separate the CI pipeline into two stages: build and test. Split off the
tests as a separate job at the 'test' stage. Replace 'build-and-run.sh'
with two additional jobs: one building an executable with lib2geom as a
subproject (build stage) and another one which finds lib2geom via cmake
and links against it (test stage, reusing the installation in the arti-
facts from the build stage). The test job retains the logs as artifacts
whenever any of the tests fail.