Better detect if a Bezier curve is a line segment
The isLineSegment() method now returns true when called on a Bézier
curve which doesn't have control points distinct from the first and
last one. This helps the curve intersection algorithm handle curves
such as CubicBezier(a, a, b, b) and treat them as segments, instead
of feeding them into the Bezier clipping algorithm. A regression
test is added for the fixed issue, based on the issue report.

Fixes: https://gitlab.com/inkscape/lib2geom/-/issues/33