Fix a logic bug in Path::appendPortionTo()
The function Path::appendPortionTo() cuts out a path portion between the
specified times.  When called with the cross_start argument set to true,
the expectation is that this portion will straddle the starting point of
the closed path. However, due to a logic bug, this wasn't happening when
the initial time of the portion was the very end of the path (i.e., time
1.0 on the last curve) or the very beginning (t=0.0 on the first curve).

To fix the bug, we add a check for whether normalizing the time forwards
or backwards swaps the sort order. When this happens and cross_start was
true, we have effectively (logically) "crossed" the start by normalizing
the time across it and no further start crossing should occur.

A test is added for all four cases to ensure that the correct portion of
the path is appended to the output.