EllipticalArc: handle the case of the same ellipse
The intersect() function for two elliptical arcs can now handle the
situation where the arcs lie on the same ellipse and report them as
intersecting. In this case, the returned intersections are the ends
of the overlapping portion between the arcs. If there is no overlap
between the arcs, no intersections are returned.

Additionally, the ellipse-to-ellipse intersection function throws a
newly added `InfinitelyManySolutions` exception when it is asked to
find intersections between two identical ellipses.