Path self-intersection: fix juncture filtering
Junctures between adjacent curves in a Path should not be considered as
self-intersections of the path and need to be filtered out. This commit
ensures that this filtering happens independently for both junctures in
a 2-curve closed path, fixing a regression from the reference.

Reference: 88040ea2aeab8ccec2b0e96c7bda2fc7d500d5ec