Division-free decomposition of degenerate conics
Add an alternative algorithm for decomposing a degenerate conic section
into lines without using floating point division. The algorithm is more
aware of numerical precision issues and uses an epsilon parameter to
detect solutions to quadratic equations in which the discriminant is
zero mathematically, but was computed as a small negative value due to
rounding errors.