Improve stability of ellipse-ellipse intersections
Rework the ellipse-to-ellipse intersection function, taking advantage
of the division-free decomposition of degenerate conics. Additionally,
the new version can handle a tangency between ellipses, even if it
appears as a single intersection on one ellipse and two intersections
on the other (due to rounding errors). New, previously failing tests
are added for the case of tangent ellipses.