Intersect low-degree Béziers and segments exactly
Provide template specializations for intersecting a quadratic or cubic
Bézier curve with a line segment. Both specializations use an algebraic
method to find the intersections: first, transform the curve to a new
coordinate system in which the segment lies on the x-axis, then find the
roots of the y-coordinate of the transformed curve using the algebraic
root finding functions available for quadratic and cubic polynomials.
Since this method is not iterative, the precision parameter is not used.

Fixes https://gitlab.com/inkscape/lib2geom/-/issues/47