Early returns from Curve::portion(0, 1)
Provide early returns in implementations of Curve::portion() for non-
trivial curve types in the cases where the requested curve portion is
either the entire curve or a reversed copy of it. In the context of
boolean operations, where long paths may be portioned, this prevents
expensive recalculations of the geometry of intermediate curves which
are included by Path::portion() in their entirety.

Since the endpoints of EllipticalArcs are no longer recalculated when
portion(0, 1) or portion(1, 0) is requested, the occasional numerical
errors in such recalculations will no longer throw ContinuityErrors.

Fixes https://gitlab.com/inkscape/lib2geom/-/issues/50