Fix intersection times for degenerate ellipses
When intersecting a degenerate ellipse or an elliptical arc with
another ellipse (possibly degenerate) or a line, provide correct
intersection time by converting from the curve time on the major
axis (onto which the degenerate ellipse collapses) to the angle
coordinate on that ellipse. Additionally, these values should be
swapped due to how ShapeIntersection is structured. Add detailed
tests for the aforementioned cases.