Improve expandToTransformed() for elliptical arcs
Add a specialised implementation of EllipticalArc::expandToTransformed()
which computes the result directly rather than transforming and taking
the bounds. The new method is less computationally-intensive, avoids
trig calculations for full arcs, and avoids any calculation if the box
to expand is already contained in the boundsFast(). In addition, a test
is written for the new code, and incorrect exception-catching is fixed
in the elliptical-arc toy.