Improve arc/segment endpoint crossing detection
The intersections of an ellipse with a line segment are found by first
intersecting the ellipse with the complete straight line and then only
keeping the intersections with the time parameter in the unit interval.

This commit ensures that this 'unit' interval is enlarged a little, so
as to account for numerical errors in the intersection routines, which
in turn improves intersection detection in the case when the endpoints
of an arc and a line segment coincide, a situation highly relevant for
boolean operations.