Report overlapping segments as intersecting
Ensure that when two line segments overlap completely (because they are
either identical or one is a reversed copy of the other), both of their
endpoints are returned as intersections between them. Furthermore, make
sure that when segments overlap only partially, they are still reported
as intersecting (either at both endpoints of the overlapping portion or
at a single point, in case numerical errors prevent the exact detection
of their collinearity).

Co-authored-by: Thomas Holder <thomas@thomas-holder.de>